flask_sync/requirements.txt
Flask~=3.0.3
mysql-connector-python~=9.1.0
elasticsearch~=8.15.1
APScheduler
Flask-APScheduler
aiomysql~=0.2.0